feat: additional tests, move tests to non-generated file
